Introduce squad/ package: one sub-package per agent role
Each squad member is now a self-contained package under squad/<role>/:
  __init__.py  — @tool functions + SquadMember subclass with build_agent()
  prompt.md    — task description and expected_output (moved from prompts/)

squad/__init__.py provides the SquadMember ABC:
  - slug (ClassVar[str]) — dict key for agents/tasks wiring
  - build_agent() — abstract, implemented per-member
  - load_prompt() — reads prompt.md via inspect.getfile(), calls _parse_prompt()
  - build_task()  — constructs the CrewAI Task from the prompt + agent

agents.py shrinks to a thin assembler: one LLM instantiation, a
_SQUAD_MEMBERS list, and a dict comprehension over slugs.

tasks.py now delegates prompt loading and Task construction to each
SquadMember class; only context chaining lives here.

Adding a new squad member = add one directory with __init__.py + prompt.md.
