+++ squad/osint_analyst/prompt.md
