+++ squad/penetration_tester/prompt.md
