+++ squad/programme_manager/prompt.md
