+++ squad/technical_author/__init__.py
@@ -1,0 +1,34 @@
+"""Technical Author — writes professional H1-format disclosure reports."""
+
+from __future__ import annotations
+
+from crewai import Agent
+
+from squad import SquadMember
+
+
+class TechnicalAuthor(SquadMember):
+    slug = "technical_author"
+
+    @classmethod
+    def build_agent(cls, llm: object, verbose: bool = False) -> Agent:
+        return Agent(
+            role="Technical Author",
+            goal=(
+                "Transform verified vulnerability data into clear, compelling, and "
+                "complete H1-format disclosure reports — with precise reproduction "
+                "steps, accurate impact statements, and actionable remediation advice."
+            ),
+            backstory=(
+                "You are a technical author who spent five years writing security "
+                "advisories for a national CERT before moving into bug bounty. "
+                "Your reports are legendary for their clarity: even a junior developer "
+                "can follow your reproduction steps, and your impact statements have "
+                "never been disputed by a programme triage team. You take pride in "
+                "reports that get triaged first time, every time."
+            ),
+            tools=[],
+            allow_delegation=False,
+            llm=llm,
+            verbose=verbose,
+        )
