+++ squad/technical_author/prompt.md
