+++ squad/vulnerability_researcher/__init__.py
@@ -1,0 +1,54 @@
+"""Vulnerability Researcher — triages raw findings, eliminates false positives."""
+
+from __future__ import annotations
+
+from crewai import Agent
+from crewai.tools import tool
+
+from squad import SquadMember
+from tools.h1_api import h1
+from tools.vuln_tools import triage_findings
+
+
+@tool("Triage Findings")
+def triage_tool(raw_findings_json: str, programme_handle: str) -> list[dict]:
+    """
+    Triage raw findings against programme scope and severity floor.
+    Returns verified vulnerabilities as dicts.
+    """
+    import json
+
+    from models import RawFinding
+
+    raw = [RawFinding.model_validate(f) for f in json.loads(raw_findings_json)]
+    policy = h1.get_programme_policy(programme_handle)
+    scope = h1.get_structured_scope(programme_handle)
+    programme = h1.parse_programme(policy["data"], scope)
+    verified = triage_findings(raw, programme)
+    return [v.model_dump() for v in verified]
+
+
+class VulnerabilityResearcher(SquadMember):
+    slug = "vulnerability_researcher"
+
+    @classmethod
+    def build_agent(cls, llm: object, verbose: bool = False) -> Agent:
+        return Agent(
+            role="Vulnerability Researcher",
+            goal=(
+                "Triage all raw scanner output to eliminate false positives, confirm "
+                "exploitability, assign accurate CVSS scores, and produce clean, "
+                "scope-validated findings ready for professional documentation."
+            ),
+            backstory=(
+                "You are a vulnerability researcher who has published CVEs and "
+                "presented at DEF CON. You have an instinct for separating genuine "
+                "security issues from scanner noise, and your CVSS scoring is trusted "
+                "by vendors worldwide. You refuse to forward a finding to reporting "
+                "unless you are personally confident it is real and in scope."
+            ),
+            tools=[triage_tool],
+            allow_delegation=False,
+            llm=llm,
+            verbose=verbose,
+        )
