+++ squad/vulnerability_researcher/prompt.md
